generate/requirements.txt
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
